Do not try to convert from ascii to utf8
ASCII is compatible with UTF8 so we need no conversion.
If ASCII detection was bogus then we hope the file is really UTF8
already, in which case we loose NO characters. If it's something
else we'll just loose the character (as it is already the case).

Basically this change fixes the "misdetected UTF8" in a better way.

See #904 for further details